refactor: group components into /search and /pokemons subfolders
