+++ src/components/index.ts
@@ -3,9 +3,9 @@ export { default as ErrorBoundary } from './ErrorBoundary';
 export { default as Footer } from './Footer';
 export { default as Loader } from './Loader';
 export { default as Pagination } from './Pagination';
-export { default as Pokemon } from './Pokemon';
-export { default as PokemonDetails } from './PokemonDetails';
-export { default as PokemonList } from './PokemonList';
-export { default as SearchPanel } from './SearchPanel';
-export { default as SearchForm } from './SearchForm';
-export { default as SearchResultsHeader } from './SearchResultsHeader';
+export { default as Pokemon } from './pokemons/Pokemon';
+export { default as PokemonDetails } from './pokemons/PokemonDetails';
+export { default as PokemonList } from './pokemons/PokemonList';
+export { default as SearchPanel } from './search/SearchPanel';
+export { default as SearchForm } from './search/SearchForm';
+export { default as SearchResultsHeader } from './search/SearchResultsHeader';
