+++ src/components/search/SearchResultsHeader.tsx
